backend/requirements.txt
Flask==3.0.3
Flask-WTF==1.2.1
WTForms==3.1.2
Werkzeug==3.0.3
pandas==2.2.2
openai==1.30.0
python-dotenv==1.0.1
Flask-Limiter==3.7.0
waitress==3.0.0
bleach==6.1.0
weasyprint==59.0
pydyf<0.11.0
bleach[css]==6.1.0
requests==2.32.2
openpyxl==3.1.2
cryptography==42.0.7
flask-talisman==1.1.0
beautifulsoup4==4.12.3
Flask-SQLAlchemy==3.1.1
huey==2.5.0
